maint: clean up bunmigrator and related docs (#11)
Followup to #8, which introduced the new `bunmigrator` package.

- Updates the `bunmigrator` documentation to show how to include the
postgres driver being used in the tests.
- Moves table-initialization logic out of `bunmigrator.Prepare` and into
`bunmigrator.Migrate`. This functions the same way, and will make my
life slightly easier as I think about removing or replacing the
`Prepare` method. (I'm thinking of removing it since it's unnecessary
and unused by any of the migrators.)
- Updates the main README to mention that `bunmigrator` exists, and
thank @BrynBerkeley for contributing it.