feat: add Config.ForceTerminateConnections for test cleanup (#13)
@jdpedrie initially wrote in in #9:

> I've found the cleanup step to cause more issues than any other part
of this (extremely useful!) library. We use it in combination with a
testcontainer instance of postgres, so it's not strictly necessary for
us, as the cleanup is handled by tearing down the docker container.

This PR adds the `ForceTerminateConnections` option to
`pgtestdb.Config`. For backwards compatibility this behavior defaults to
`false`. When set to `true`, pgtestdb will force-close any connections
to the testdb instance before attempting to drop the database.

This PR confirms that this behavior works as expected by adding unit
tests that confirm (a) an open connection causes a test failure when
`ForceTerminateConnections == false`, and (b) an open connection is
terminated and the database is dropped successfully when
`ForceTerminateConnections == true`.

When a test fails due to hanging/leaked/open connections, and
`ForceTerminateConnections == false`, pgtestdb will log a message
explaining the situation and suggesting setting it to `true` instead:

```
--- FAIL: TestDefaultDroppingBehaviorFailsWithOpenConnections (5.36s)
    /Users/pd/code/pgtestdb/testdb_test.go:373: testdbconf: postgres://pgtdbuser:pgtdbpass@localhost:5433/testdb_tpl_0b52ddb8de835787ddfb243752fe438f_inst_934fb071?sslmode=disable
    /Users/pd/code/pgtestdb/testdb.go:265: pgtestdb failed to clean up the test database because there are still open connections to it.
    /Users/pd/code/pgtestdb/testdb.go:266: This usually means that your code is leaking database connections, which is usually bad.
    /Users/pd/code/pgtestdb/testdb.go:267: If you would like pgtestdb to force-terminate any open connections at the end of the testcase, set `ForceTerminateConnections = true` on your `pgtestdb.Config`
    /Users/pd/code/pgtestdb/testdb.go:269: could not drop test database 'testdb_tpl_0b52ddb8de835787ddfb243752fe438f_inst_934fb071': ERROR: database "testdb_tpl_0b52ddb8de835787ddfb243752fe438f_inst_934fb071" is being accessed by other users (SQLSTATE 55006)
```

After this PR is merged, pgtestdb will need to be version-bumped and
tagged. After that, the migrators will need to be upgraded to depend on
this new version of pgtestdb, version-bumped, and tagged.

---------

Co-authored-by: Peter Downs <github@peterdowns.com>