docs: explain connecting to unix domain sockets /var/postgresql (#32)
Fixes https://github.com/peterldowns/pgtestdb/issues/27 by adding a FAQ
entry to the README explaining how to connect to a server over unix
domain socket.

Although [the Postgres
documentation](https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-CONNECT-HOST)
explains that you can use a `?host=%2Fvar%2Fpostgres` connection string
option instead of specifying the host in the URL, I didn't know this and
I imagine most people also are unfamiliar with this option.

And, although the documentation says that you can pass a URL-encoded
host in the connection string and just expect it to work, that doesn't
happen with `sql.Open` — it always fails.

So, thanks @ptman for opening the issue and discovering the solution,
this PR just adds that information to the README.